requirements.txt
yarl==1.8.2
fronzenlist==1.3.3
multidict==6.0.4
requests==2.28.2
PyNaCl==1.5.0
discord==2.1.0
python-dotenv==0.21.1
discord-pretty-help==2.0.0
Pillow==9.4.0
apscheduler==3.9.1